archive/phase1/pyproject.toml
[project]
name = "atom"
version = "0.1.0"
description = "Video-based human motion understanding platform"
requires-python = ">=3.10"
dependencies = [
    "mediapipe>=0.10",
    "opencv-python>=4.8",
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
ml = [
    "scikit-learn>=1.3",
    "torch>=2.0",
    "pyarrow>=14.0",
]
server = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "google-cloud-storage>=2.18",
    "pydantic-settings>=2.6",
    "python-multipart>=0.0.18",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
